move intersects to window-utils
