+++ src/node/desktop/src/main/window-utils.ts
@@ -13,7 +13,7 @@
  *
  */
 
-import { BrowserWindow, webContents, WebContents } from 'electron';
+import { screen, BrowserWindow, webContents, WebContents, Rectangle } from 'electron';
 import { appState } from './app-state';
 import { PendingSatelliteWindow, PendingSecondaryWindow } from './pending-window';
 import { SatelliteWindow } from './satellite-window';
@@ -134,32 +134,63 @@ export interface WindowBounds {
 }
 
 /**
+ * Check if two Rectangles intersect.
+ *
+ * @param one First rectangle
+ * @param two Second rectangle
+ *
+ * @returns True if at least one pixel is within both rectangles.
+ */
+export function intersects(first: Rectangle, second: Rectangle): boolean {
+  const firstRight = first.x + first.width;
+  const firstBottom = first.y + first.height;
+  const secondRight = second.x + second.width;
+  const secondBottom = second.y + second.height;
+
+  if (first.x >= secondRight || firstRight <= second.x) {
+    // Don't overlap horizontally
+    return false;
+  }
+
+  if (first.y >= secondBottom || firstBottom <= second.y) {
+    // Don't overlap vertically
+    return false;
+  }
+
+  return true;
+}
+
+/**
  * Resize and reposition a window to requested location, and ensure it is visible and not too
  * small (e.g. if monitor configuration has changed).
- * 
- * @param window window to 
+ *
+ * @param window window to reposition and resize
+ * @param requestedBounds requested size and position
  * @param defaultWidth width to use if width must be changed
  * @param defaultHeight height to use if height must be changed
  */
-// export function positionAndEnsureVisible(window: BrowserWindow, defaultWidth: number, defaultHeight: number) {
-//   const savedBounds = this.windowBounds();
-  
+// export function positionAndEnsureVisible(
+//   window: BrowserWindow,
+//   requestedBounds: WindowBounds,
+//   defaultWidth: number,
+//   defaultHeight: number) {
+
 //   // Check if saved bounds is still partially in one of the available displays.
-  
+
 //   // Shrink the window rectangle a bit just to capture cases like RStudio
 //   // too close to edge of window and hardly showing at all.
 //   const checkRect = {
-//     height: savedBounds.height - 5,
-//     width: savedBounds.width - 5,
-//     x: savedBounds.x + 5,
-//     y: savedBounds.y + 5,
+//     height: requestedBounds.height - 5,
+//     width: requestedBounds.width - 5,
+//     x: requestedBounds.x + 5,
+//     y: requestedBounds.y + 5,
 //   };
-  
+
 //   // check for intersection
 //   const goodDisplays = screen.getAllDisplays().find((display) => {
 //     return intersects(checkRect, display.workArea);
 //   });
-  
+
 //   // Restore it to previous location if possible, or center of primary display otherwise
 //   if (goodDisplays) {
 //     mainWindow.setBounds(savedBounds);
@@ -169,9 +200,9 @@ export interface WindowBounds {
 //       width: Math.min(defaultWidth, primaryBounds.width),
 //       height: Math.min(defaultHeight, primaryBounds.height),
 //     };
-    
+
 //     mainWindow.setSize(newSize.width, newSize.height);
-    
+
 //     // window.center() doesn't consistently pick the primary display,
 //     // so manually calculating the center of the primary display
 //     mainWindow.setPosition(
@@ -179,9 +210,9 @@ export interface WindowBounds {
 //       primaryBounds.y + (primaryBounds.height - newSize.height) / 2,
 //       );
 //     }
-    
+
 //     // ensure a minimum size for the window on restore
 //     const currSize = mainWindow.getSize();
 //     mainWindow.setSize(Math.max(300, currSize[0]), Math.max(200, currSize[1]));
-    
+
 // }
